+++ include/gtirb_pprinter/Assembler.hpp
@@ -25,8 +25,8 @@ class DEBLOAT_PRETTYPRINTER_EXPORT_API Assembler {
 public:
   virtual ~Assembler() = default;
 
-  virtual std::string cleanByte(uint8_t b) const;
-  virtual std::string cleanSingleQuote(const std::string& s) const;
+  virtual std::string escapeByte(uint8_t b) const;
+  virtual std::string escapeSingleQuote(const std::string& s) const;
 };
 
 class DEBLOAT_PRETTYPRINTER_EXPORT_API GasAssembler : public Assembler {};